Struct aws_sdk_location::client::fluent_builders::CalculateRoute

pub struct CalculateRoute<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CalculateRoute<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CalculateRouteOutput, SdkError<CalculateRouteError>> where
R::Policy: SmithyRetryPolicy<CalculateRouteInputOperationOutputAlias, CalculateRouteOutput, CalculateRouteError, CalculateRouteInputOperationRetryAlias>,

pub fn calculator_name(self, inp: impl Into<String>) -> Self

The name of the route calculator resource that you want to use to calculate a route.

pub fn set_calculator_name(self, input: Option<String>) -> Self

pub fn departure_position(self, inp: impl Into<f64>) -> Self

Appends an item to DeparturePosition.

To override the contents of this collection use set_departure_position.

The start position for the route. Defined in WGS 84 format: [longitude, latitude].

For example, [-123.115, 49.285]

If you specify a departure that's not located on a road, Amazon Location moves the position to the nearest road.

Valid Values: [-180 to 180,-90 to 90]

pub fn set_departure_position(self, input: Option<Vec<f64>>) -> Self

pub fn destination_position(self, inp: impl Into<f64>) -> Self

Appends an item to DestinationPosition.

To override the contents of this collection use set_destination_position.

The finish position for the route. Defined in WGS 84 format: [longitude, latitude].

For example, [-122.339, 47.615]

If you specify a destination that's not located on a road, Amazon Location moves the position to the nearest road.

Valid Values: [-180 to 180,-90 to 90]

pub fn set_destination_position(self, input: Option<Vec<f64>>) -> Self

pub fn waypoint_positions(self, inp: impl Into<Vec<f64>>) -> Self

Appends an item to WaypointPositions.

To override the contents of this collection use set_waypoint_positions.

Specifies an ordered list of up to 23 intermediate positions to include along a route between the departure position and destination position.

For example, from the DeparturePosition [-123.115, 49.285], the route follows the order that the waypoint positions are given [[-122.757, 49.0021],[-122.349, 47.620]]

If you specify a waypoint position that's not located on a road, Amazon Location moves the position to the nearest road.

Specifying more than 23 waypoints returns a 400 ValidationException error.

Valid Values: [-180 to 180,-90 to 90]

pub fn set_waypoint_positions(self, input: Option<Vec<Vec<f64>>>) -> Self

pub fn travel_mode(self, inp: TravelMode) -> Self

Specifies the mode of transport when calculating a route. Used in estimating the speed of travel and road compatibility.

The TravelMode you specify determines how you specify route preferences:

If traveling by Car use the CarModeOptions parameter.

If traveling by Truck use the TruckModeOptions parameter.

Default Value: Car

pub fn set_travel_mode(self, input: Option<TravelMode>) -> Self

pub fn departure_time(self, inp: Instant) -> Self

Specifies the desired time of departure. Uses the given time to calculate a route. Otherwise, the best time of day to travel with the best traffic conditions is used to calculate the route.

Setting a departure time in the past returns a 400 ValidationException error.

In ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. For example, 2020–07-2T12:15:20.000Z+01:00

pub fn set_departure_time(self, input: Option<Instant>) -> Self

pub fn depart_now(self, inp: bool) -> Self

Sets the time of departure as the current time. Uses the current time to calculate a route. Otherwise, the best time of day to travel with the best traffic conditions is used to calculate the route.

Default Value: false

Valid Values: false | true

pub fn set_depart_now(self, input: Option<bool>) -> Self

pub fn distance_unit(self, inp: DistanceUnit) -> Self

Set the unit system to specify the distance.

Default Value: Kilometers

pub fn set_distance_unit(self, input: Option<DistanceUnit>) -> Self

pub fn include_leg_geometry(self, inp: bool) -> Self

Set to include the geometry details in the result for each path between a pair of positions.

Default Value: false

Valid Values: false | true

pub fn set_include_leg_geometry(self, input: Option<bool>) -> Self

pub fn car_mode_options(self, inp: CalculateRouteCarModeOptions) -> Self

Specifies route preferences when traveling by Car, such as avoiding routes that use ferries or tolls.

Requirements: TravelMode must be specified as Car.

pub fn set_car_mode_options(
self,
input: Option<CalculateRouteCarModeOptions>
) -> Self

pub fn truck_mode_options(self, inp: CalculateRouteTruckModeOptions) -> Self

Specifies route preferences when traveling by Truck, such as avoiding routes that use ferries or tolls, and truck specifications to consider when choosing an optimal road.

Requirements: TravelMode must be specified as Truck.

pub fn set_truck_mode_options(
self,
input: Option<CalculateRouteTruckModeOptions>
) -> Self