Struct aws_sdk_location::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for LocationService.

This client allows ergonomic access to a LocationService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_tracker_consumer(&self) -> AssociateTrackerConsumer<C, M, R>

pub fn batch_delete_device_position_history(
&self
) -> BatchDeleteDevicePositionHistory<C, M, R>

pub fn batch_delete_geofence(&self) -> BatchDeleteGeofence<C, M, R>

pub fn batch_evaluate_geofences(&self) -> BatchEvaluateGeofences<C, M, R>

pub fn batch_get_device_position(&self) -> BatchGetDevicePosition<C, M, R>

pub fn batch_put_geofence(&self) -> BatchPutGeofence<C, M, R>

pub fn batch_update_device_position(&self) -> BatchUpdateDevicePosition<C, M, R>

pub fn calculate_route(&self) -> CalculateRoute<C, M, R>

pub fn create_geofence_collection(&self) -> CreateGeofenceCollection<C, M, R>

pub fn create_map(&self) -> CreateMap<C, M, R>

pub fn create_place_index(&self) -> CreatePlaceIndex<C, M, R>

pub fn create_route_calculator(&self) -> CreateRouteCalculator<C, M, R>

pub fn create_tracker(&self) -> CreateTracker<C, M, R>

pub fn delete_geofence_collection(&self) -> DeleteGeofenceCollection<C, M, R>

pub fn delete_map(&self) -> DeleteMap<C, M, R>

pub fn delete_place_index(&self) -> DeletePlaceIndex<C, M, R>

pub fn delete_route_calculator(&self) -> DeleteRouteCalculator<C, M, R>

pub fn delete_tracker(&self) -> DeleteTracker<C, M, R>

pub fn describe_geofence_collection(
&self
) -> DescribeGeofenceCollection<C, M, R>

pub fn describe_map(&self) -> DescribeMap<C, M, R>

pub fn describe_place_index(&self) -> DescribePlaceIndex<C, M, R>

pub fn describe_route_calculator(&self) -> DescribeRouteCalculator<C, M, R>

pub fn describe_tracker(&self) -> DescribeTracker<C, M, R>

pub fn disassociate_tracker_consumer(
&self
) -> DisassociateTrackerConsumer<C, M, R>

pub fn get_device_position(&self) -> GetDevicePosition<C, M, R>

pub fn get_device_position_history(&self) -> GetDevicePositionHistory<C, M, R>

pub fn get_geofence(&self) -> GetGeofence<C, M, R>

pub fn get_map_glyphs(&self) -> GetMapGlyphs<C, M, R>

pub fn get_map_sprites(&self) -> GetMapSprites<C, M, R>

pub fn get_map_style_descriptor(&self) -> GetMapStyleDescriptor<C, M, R>

pub fn get_map_tile(&self) -> GetMapTile<C, M, R>

pub fn list_device_positions(&self) -> ListDevicePositions<C, M, R>

pub fn list_geofence_collections(&self) -> ListGeofenceCollections<C, M, R>

pub fn list_geofences(&self) -> ListGeofences<C, M, R>

pub fn list_maps(&self) -> ListMaps<C, M, R>

pub fn list_place_indexes(&self) -> ListPlaceIndexes<C, M, R>

pub fn list_route_calculators(&self) -> ListRouteCalculators<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_tracker_consumers(&self) -> ListTrackerConsumers<C, M, R>

pub fn list_trackers(&self) -> ListTrackers<C, M, R>

pub fn put_geofence(&self) -> PutGeofence<C, M, R>

pub fn search_place_index_for_position(
&self
) -> SearchPlaceIndexForPosition<C, M, R>

pub fn search_place_index_for_text(&self) -> SearchPlaceIndexForText<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_geofence_collection(&self) -> UpdateGeofenceCollection<C, M, R>

pub fn update_map(&self) -> UpdateMap<C, M, R>

pub fn update_place_index(&self) -> UpdatePlaceIndex<C, M, R>

pub fn update_route_calculator(&self) -> UpdateRouteCalculator<C, M, R>

pub fn update_tracker(&self) -> UpdateTracker<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self