Module aws_sdk_location::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

The request was denied because of insufficient access or permissions. Check with an

AssociateTrackerConsumerError

BatchDeleteDevicePositionHistoryError

BatchDeleteGeofenceError

BatchEvaluateGeofencesError

BatchGetDevicePositionError

BatchPutGeofenceError

BatchUpdateDevicePositionError

CalculateRouteError

ConflictException

The request was unsuccessful because of a conflict.

CreateGeofenceCollectionError

CreateMapError

CreatePlaceIndexError

CreateRouteCalculatorError

CreateTrackerError

DeleteGeofenceCollectionError

DeleteMapError

DeletePlaceIndexError

DeleteRouteCalculatorError

DeleteTrackerError

DescribeGeofenceCollectionError

DescribeMapError

DescribePlaceIndexError

DescribeRouteCalculatorError

DescribeTrackerError

DisassociateTrackerConsumerError

GetDevicePositionError

GetDevicePositionHistoryError

GetGeofenceError

GetMapGlyphsError

GetMapSpritesError

GetMapStyleDescriptorError

GetMapTileError

InternalServerException

The request has failed to process because of an unknown server error, exception, or failure.

ListDevicePositionsError

ListGeofenceCollectionsError

ListGeofencesError

ListMapsError

ListPlaceIndexesError

ListRouteCalculatorsError

ListTagsForResourceError

ListTrackerConsumersError

ListTrackersError

PutGeofenceError

ResourceNotFoundException

The resource that you've entered was not found in your AWS account.

SearchPlaceIndexForPositionError

SearchPlaceIndexForTextError

ServiceQuotaExceededException

The operation was denied because the request would exceed the maximum quota

TagResourceError

ThrottlingException

The request was denied because of request throttling.

UntagResourceError

UpdateGeofenceCollectionError

UpdateMapError

UpdatePlaceIndexError

UpdateRouteCalculatorError

UpdateTrackerError

ValidationException

The input failed to meet the constraints specified by the AWS service.

Enums

AssociateTrackerConsumerErrorKind

BatchDeleteDevicePositionHistoryErrorKind

BatchDeleteGeofenceErrorKind

BatchEvaluateGeofencesErrorKind

BatchGetDevicePositionErrorKind

BatchPutGeofenceErrorKind

BatchUpdateDevicePositionErrorKind

CalculateRouteErrorKind

CreateGeofenceCollectionErrorKind

CreateMapErrorKind

CreatePlaceIndexErrorKind

CreateRouteCalculatorErrorKind

CreateTrackerErrorKind

DeleteGeofenceCollectionErrorKind

DeleteMapErrorKind

DeletePlaceIndexErrorKind

DeleteRouteCalculatorErrorKind

DeleteTrackerErrorKind

DescribeGeofenceCollectionErrorKind

DescribeMapErrorKind

DescribePlaceIndexErrorKind

DescribeRouteCalculatorErrorKind

DescribeTrackerErrorKind

DisassociateTrackerConsumerErrorKind

GetDevicePositionErrorKind

GetDevicePositionHistoryErrorKind

GetGeofenceErrorKind

GetMapGlyphsErrorKind

GetMapSpritesErrorKind

GetMapStyleDescriptorErrorKind

GetMapTileErrorKind

ListDevicePositionsErrorKind

ListGeofenceCollectionsErrorKind

ListGeofencesErrorKind

ListMapsErrorKind

ListPlaceIndexesErrorKind

ListRouteCalculatorsErrorKind

ListTagsForResourceErrorKind

ListTrackerConsumersErrorKind

ListTrackersErrorKind

PutGeofenceErrorKind

SearchPlaceIndexForPositionErrorKind

SearchPlaceIndexForTextErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateGeofenceCollectionErrorKind

UpdateMapErrorKind

UpdatePlaceIndexErrorKind

UpdateRouteCalculatorErrorKind

UpdateTrackerErrorKind