Module aws_sdk_location::model

Modules

batch_delete_device_position_history_error

See BatchDeleteDevicePositionHistoryError

batch_delete_geofence_error

See BatchDeleteGeofenceError

batch_evaluate_geofences_error

See BatchEvaluateGeofencesError

batch_get_device_position_error

See BatchGetDevicePositionError

batch_item_error

See BatchItemError

batch_put_geofence_error

See BatchPutGeofenceError

batch_put_geofence_request_entry

See BatchPutGeofenceRequestEntry

batch_put_geofence_success

See BatchPutGeofenceSuccess

batch_update_device_position_error

See BatchUpdateDevicePositionError

calculate_route_car_mode_options

See CalculateRouteCarModeOptions

calculate_route_summary

See CalculateRouteSummary

calculate_route_truck_mode_options

See CalculateRouteTruckModeOptions

data_source_configuration

See DataSourceConfiguration

device_position

See DevicePosition

device_position_update

See DevicePositionUpdate

geofence_geometry

See GeofenceGeometry

leg

See Leg

leg_geometry

See LegGeometry

list_device_positions_response_entry

See ListDevicePositionsResponseEntry

list_geofence_collections_response_entry

See ListGeofenceCollectionsResponseEntry

list_geofence_response_entry

See ListGeofenceResponseEntry

list_maps_response_entry

See ListMapsResponseEntry

list_place_indexes_response_entry

See ListPlaceIndexesResponseEntry

list_route_calculators_response_entry

See ListRouteCalculatorsResponseEntry

list_trackers_response_entry

See ListTrackersResponseEntry

map_configuration

See MapConfiguration

place

See Place

place_geometry

See PlaceGeometry

search_for_position_result

See SearchForPositionResult

search_for_text_result

See SearchForTextResult

search_place_index_for_position_summary

See SearchPlaceIndexForPositionSummary

search_place_index_for_text_summary

See SearchPlaceIndexForTextSummary

step

See Step

truck_dimensions

See TruckDimensions

truck_weight

See TruckWeight

validation_exception_field

See ValidationExceptionField

Structs

BatchDeleteDevicePositionHistoryError

Contains the tracker resource details.

BatchDeleteGeofenceError

Contains error details for each geofence that failed to delete from the geofence

BatchEvaluateGeofencesError

Contains error details for each device that failed to evaluate its position against

BatchGetDevicePositionError

Contains error details for each device that didn't return a position.

BatchItemError

Contains the batch request error details associated with the request.

BatchItemErrorCode

BatchPutGeofenceError

Contains error details for each geofence that failed to be stored in a given geofence

BatchPutGeofenceRequestEntry

Contains geofence geometry details.

BatchPutGeofenceSuccess

Contains a summary of each geofence that was successfully stored in a given geofence

BatchUpdateDevicePositionError

Contains error details for each device that failed to update its position.

CalculateRouteCarModeOptions

Contains details about additional route preferences for requests that specify

CalculateRouteSummary

A summary of the calculated route.

CalculateRouteTruckModeOptions

Contains details about additional route preferences for requests that specify

DataSourceConfiguration

Specifies the data storage option chosen for requesting Places.

DevicePosition

Contains the device position details.

DevicePositionUpdate

Contains the position update details for a device.

DimensionUnit

DistanceUnit

GeofenceGeometry

Contains the geofence geometry details.

IntendedUse

Leg

Contains the calculated route's details for each path between a pair of positions. The

LegGeometry

Contains the geometry details for each path between a pair of positions. Used in

ListDevicePositionsResponseEntry

Contains the tracker resource details.

ListGeofenceCollectionsResponseEntry

Contains the geofence collection details.

ListGeofenceResponseEntry

Contains a list of geofences stored in a given geofence collection.

ListMapsResponseEntry

Contains details of an existing map resource in your AWS account.

ListPlaceIndexesResponseEntry

A place index resource listed in your AWS account.

ListRouteCalculatorsResponseEntry

A route calculator resource listed in your AWS account.

ListTrackersResponseEntry

Contains the tracker resource details.

MapConfiguration

Specifies the map tile style selected from an available provider.

Place

Contains details about addresses or points of interest that match the search

PlaceGeometry

Places uses a point geometry to specify a location or a Place.

PricingPlan

SearchForPositionResult

Specifies a single point of interest, or Place as a result of a search query obtained

SearchForTextResult

Contains relevant Places returned by calling

SearchPlaceIndexForPositionSummary

A summary of the reverse geocoding request sent using SearchPlaceIndexForPosition.

SearchPlaceIndexForTextSummary

A summary of the geocoding request sent using SearchPlaceIndexForText.

Step

Represents an element of a leg within a route. A step contains instructions for how

TravelMode

TruckDimensions

Contains details about the truck dimensions in the unit of measurement that you

TruckWeight

Contains details about the truck's weight specifications. Used to avoid roads that

ValidationExceptionField

The input failed to meet the constraints specified by the AWS service in a specified

ValidationExceptionReason

VehicleWeightUnit