Module aws_sdk_location::operation

Structs

AssociateTrackerConsumer

Creates an association between a geofence collection and a tracker resource. This

BatchDeleteDevicePositionHistory

Deletes the position history of one or more devices from a tracker resource.

BatchDeleteGeofence

Deletes a batch of geofences from a geofence collection.

BatchEvaluateGeofences

Evaluates device positions against the geofence geometries from a given geofence

BatchGetDevicePosition

Lists the latest device positions for requested devices.

BatchPutGeofence

A batch request for storing geofence geometries into a given geofence collection, or

BatchUpdateDevicePosition

Uploads position update data for one or more devices to a tracker resource. Amazon Location

CalculateRoute

CreateGeofenceCollection

Creates a geofence collection, which manages and stores geofences.

CreateMap

Creates a map resource in your AWS account, which provides map tiles of different

CreatePlaceIndex

Creates a place index resource in your AWS account, which supports functions with

CreateRouteCalculator

Creates a route calculator resource in your AWS account.

CreateTracker

Creates a tracker resource in your AWS account, which lets you retrieve current and

DeleteGeofenceCollection

Deletes a geofence collection from your AWS account.

DeleteMap

Deletes a map resource from your AWS account.

DeletePlaceIndex

Deletes a place index resource from your AWS account.

DeleteRouteCalculator

Deletes a route calculator resource from your AWS account.

DeleteTracker

Deletes a tracker resource from your AWS account.

DescribeGeofenceCollection

Retrieves the geofence collection details.

DescribeMap

Retrieves the map resource details.

DescribePlaceIndex

Retrieves the place index resource details.

DescribeRouteCalculator

Retrieves the route calculator resource details.

DescribeTracker

Retrieves the tracker resource details.

DisassociateTrackerConsumer

Removes the association between a tracker resource and a geofence collection.

GetDevicePosition

Retrieves a device's most recent position according to its sample time.

GetDevicePositionHistory

Retrieves the device position history from a tracker resource within a specified range

GetGeofence

Retrieves the geofence details from a geofence collection.

GetMapGlyphs

Retrieves glyphs used to display labels on a map.

GetMapSprites

Retrieves the sprite sheet corresponding to a map resource. The sprite sheet is a PNG

GetMapStyleDescriptor

Retrieves the map style descriptor from a map resource.

GetMapTile

Retrieves a vector data tile from the map resource. Map tiles are used by clients to

ListDevicePositions

A batch request to retrieve all device positions.

ListGeofenceCollections

Lists geofence collections in your AWS account.

ListGeofences

Lists geofences stored in a given geofence collection.

ListMaps

Lists map resources in your AWS account.

ListPlaceIndexes

Lists place index resources in your AWS account.

ListRouteCalculators

Lists route calculator resources in your AWS account.

ListTagsForResource

Returns a list of tags that are applied to the specified Amazon Location resource.

ListTrackerConsumers

Lists geofence collections currently associated to the given tracker resource.

ListTrackers

Lists tracker resources in your AWS account.

PutGeofence

Stores a geofence geometry in a given geofence collection, or updates the geometry of

SearchPlaceIndexForPosition

Reverse geocodes a given coordinate and returns a legible address. Allows you to search

SearchPlaceIndexForText

Geocodes free-form text, such as an address, name, city, or region to allow you to

TagResource

Assigns one or more tags (key-value pairs) to the specified Amazon

UntagResource

Removes one or more tags from the specified Amazon Location resource.

UpdateGeofenceCollection

Updates the specified properties of a given geofence collection.

UpdateMap

Updates the specified properties of a given map resource.

UpdatePlaceIndex

Updates the specified properties of a given place index resource.

UpdateRouteCalculator

Updates the specified properties for a given route calculator resource.

UpdateTracker

Updates the specified properties of a given tracker resource.